Update diskdriver for 64 bit file support.

+++ graphquery/core/models/lpg_mmap/index_file.hpp
@@ -44,7 +44,7 @@ namespace graphquery::database::storage
          ***************************************************************/
         struct SIndexEntry_t
         {
-            uint64_t id     = {};
+            int64_t id      = {};
             uint32_t offset = END_INDEX;
         };
 
@@ -60,7 +60,7 @@ namespace graphquery::database::storage
         inline SRef_t<SIndexMetadata_t> read_metadata() noexcept;
         inline SRef_t<SIndexEntry_t> read_entry(int64_t offset) noexcept;
         void open(std::filesystem::path path, std::string_view file_name, bool create) noexcept;
-        void store_entry(uint64_t id, int64_t offset) noexcept;
+        void store_entry(int64_t id, int64_t offset) noexcept;
 
       private:
         CDiskDriver m_file;
@@ -95,9 +95,9 @@ graphquery::database::storage::CIndexFile::read_entry(const int64_t offset) noex
 }
 
 inline void
-graphquery::database::storage::CIndexFile::store_entry(const uint64_t id, const int64_t offset) noexcept
+graphquery::database::storage::CIndexFile::store_entry(const int64_t id, const int64_t offset) noexcept
 {
-    auto index_ptr     = read_entry(static_cast<int64_t>(id));
+    auto index_ptr     = read_entry(id);
     const auto index_c = utils::atomic_load(&read_metadata()->index_c);
 
     utils::atomic_store(&index_ptr->id, id);
